ReactJS/FinalProject/server/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
